moved Feed.tsx and Feed.css to Feed directory. same for Navbar
