+++ front-end/src/App.tsx
@@ -1,6 +1,6 @@
 import './App.css';
-import Navbar from "./components/Navbar";
-import Feed from "./components/Feed";
+import Navbar from "./components/Navbar/Navbar";
+import Feed from "./components/Feed/Feed";
 
 function App() {
   return (
